+++ go-server/appdb_schema.sql
@@ -191,3 +191,13 @@ CREATE TABLE IF NOT EXISTS "OmniDB_app_notifychannel" (
 );
 CREATE INDEX IF NOT EXISTS "OmniDB_app_notifychannel_user_id" ON "OmniDB_app_notifychannel" ("user_id");
 CREATE INDEX IF NOT EXISTS "OmniDB_app_notifychannel_connection_id" ON "OmniDB_app_notifychannel" ("connection_id");
+
+CREATE TABLE IF NOT EXISTS "OmniDB_app_connectionorder" (
+	"id" integer NOT NULL PRIMARY KEY AUTOINCREMENT,
+	"user_id" integer NOT NULL REFERENCES "auth_user" ("id") DEFERRABLE INITIALLY DEFERRED,
+	"connection_id" bigint NOT NULL REFERENCES "OmniDB_app_connection" ("id") DEFERRABLE INITIALLY DEFERRED,
+	"position" integer NOT NULL,
+	CONSTRAINT "unique_connectionorder" UNIQUE ("user_id", "connection_id")
+);
+CREATE INDEX IF NOT EXISTS "OmniDB_app_connectionorder_user_id" ON "OmniDB_app_connectionorder" ("user_id");
+CREATE INDEX IF NOT EXISTS "OmniDB_app_connectionorder_connection_id" ON "OmniDB_app_connectionorder" ("connection_id");